+++ backend/baguni-batch/src/main/java/baguni/batch/domain/link/service/CrawlingEventListener.java
@@ -1,14 +1,14 @@
-package baguni.api.infrastructure;
+package baguni.batch.domain.link.service;
 
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.amqp.rabbit.annotation.RabbitHandler;
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.stereotype.Component;
 
-import baguni.api.service.link.dto.LinkResult;
-import baguni.api.service.link.service.LinkService;
 import baguni.common.config.RabbitmqConfig;
 import baguni.common.event.events.CrawlingEvent;
+import baguni.common.event.events.LinkEvent;
+import baguni.entity.model.link.dto.LinkResult;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -23,16 +23,30 @@ import lombok.extern.slf4j.Slf4j;
 @RabbitListener(queues = {RabbitmqConfig.QUEUE.PICK_CRAWLING})
 public class CrawlingEventListener {
 
-	private final LinkService linkService;
+	private final LinkCrawlingService linkCrawlingService;
 
 	@RabbitHandler
 	public void linkCrawlingEvent(CrawlingEvent event) {
-		LinkResult link = linkService.getLinkResult(event.getLinkId(), event.getUrl(), event.getTitle());
+		LinkResult link = linkCrawlingService.getLinkResult(event.getLinkId(), event.getUrl(), event.getTitle());
 
 		// imageUrl, description이 비어있는 경우에만 OG 태그 업데이트 시도
 		if (StringUtils.isEmpty(link.imageUrl()) || StringUtils.isEmpty(link.description())) {
 			try {
-				linkService.saveLinkAndUpdateOgTagBySelenium(link.url(), link.title());
+				linkCrawlingService.saveLinkAndUpdateOgTagBySelenium(link.url(), link.title());
+			} catch (Exception e) {
+				log.info("메세지 큐에서 꺼낸 Link OG 크롤링 실패 : ", e);
+			}
+		}
+	}
+	
+	@RabbitListener
+	public void linkUrlEvent(LinkEvent event) {
+		LinkResult link = linkCrawlingService.getLinkResultByUrl(event.getUrl());
+
+		// imageUrl, description이 비어있는 경우에만 OG 태그 업데이트 시도
+		if (StringUtils.isEmpty(link.imageUrl()) || StringUtils.isEmpty(link.description())) {
+			try {
+				linkCrawlingService.saveLinkAndUpdateOgTagBySelenium(link.url(), link.title());
 			} catch (Exception e) {
 				log.info("메세지 큐에서 꺼낸 Link OG 크롤링 실패 : ", e);
 			}
